+++ docs/Project.toml
@@ -1,0 +1,3 @@
+[deps]
+Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+Catch22 = "0b325660-c155-474b-b0c5-aee57a11d6da"
